Rename readme.md to README.md
## Big Data Analytics projects using python
### Machine learning using PySpark
This project is about learning basic of ApacheSpark. In this project, I applied most common machine learning models e.g classification, clustering, prediction analysis and NLP using PySpark.

### Machine learning project - Predictive modelling
In this projet I applied different regression algorithms to predict the price of second hand car.

### Deep learning project - Multi-class image classification using Convolutional Neural Networks (CNN)
This project is made for the Analytical Service Development course of Master of Engineering in Big Data Analytics at Arcada University of Applied Sciences, Helsinki. The aim of this work was to build a model that can classify whether an image is handgun, rifle or other using convolutional neural networks (CNN). The project also aimed to identify the best image augmentation parameter in improving the multi-class classifier or CNN model.

Dataset
The project uses weapons data from the Department of Business Management and Analytics, Arcada University of Applied Science, provided by Stagezero Oy. This is the first work using these dataset. The dataset consisted of 26,542 images from 3 classes : handgun, rifle and other.
We first started with organizing the data: splitting the data into two folders namely train and test folders. We have imported the necessary libraries for analysis. We applied Keras with a Tensorflow backend for the CNN model and the ImageDataGenerator class of Keras deep learning library for image data augmentation.

Modelling approach
The approach was to first build a simple neural network with only two hidden layers. This was used as baseline model (sequential model of Keras library). The architecture of the baseline model was as  a Conv2D layer of 32 channel of 3x3 kernel followed by flatten layer and output layer. We used  softmax activation function as we have 3 classes in the dataset. In the tuned model we added MaxPool2D layer, Conv2D of 64 channel and Dropout layer.

We observed that tuning of sequential model parameter and image augmentation technique improve the accuracy from 0.66 to 0.86 for the multi-level classification of weapons.

### Big data analysis - predicting New York taxi fare 
This project is made for the Big Data Analytics course of Master of Engineering in Big Data Analytics at Arcada Univeristy of Applied Sciences, Helsinki. The aim of this work was to work with a really big dataset – to explore the skills on machine learning using big data set one whose size would make it unfeasible to analyse and visualize using any analytical program. For this purposes, we utilized CSC super computer and puthi. The main aim was to predict the fare amount for a taxi ride in New York City, given the pick up, drop off locations and the date time of the pick up.

We downloaded multiple CSV files during 2013-2019 as URL (of 114 gigabytes) from the New York City Taxi and Limousine Commission (data source : http://www.nyc.gov/html/tlc/html/about/trip_record_data.shtml ) in puhti CSC environment using ‘wget’ command. After combining all the CSV files using ‘cat’ command we moved all the data from puhti to pouta CSC environment using allas module. Then we used valohai to analyze the data. 

A random subset of 10,0000 observations from Januray 2019 were used for building the model. First we used linear regression model as baseline model. Then we applied regularization techniques in order to solve the overfitting problem. We used sklearn’s Lasso Regression regularization in this case. We applied hyper-parameter tuning in Lasso Regression and finally applied more complex model e.g the random forest regression model in order the compare the accuracy.

In our model analysis we found LassoCV model to give better accuracy compared to other models.

### Time series analysis - Forecasting stock of Yahoo finance using time series analysis and regression model
This is the first project of machine learning model for the Master of Engineering in Big Data Analytics at Arcada Univeristy of Applied Sciences, Helsinki. The aim of this project was to predict future stock returns based on past returns of Yfinance using regression model. In this project I used a dataset containing stock prices of Yahoo from January-2014 to September-2019. Here I mainly focused on extraction of features based on the financial technical indicator from the following paper [Link of paper] (https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4873195/) and determining a decision, by forecasting 1-day ahead. 

Methodology
1.	Feature Extraction
2.	Pre-processing 
3.	Model building
4.	Model prediction
5.	Decision making of selling or buying